test_schedule_endpoint.py
import requests
import json

try:
    response = requests.get('http://127.0.0.1:8900/routines/active/schedule')
    if response.status_code == 200:
        print("Success! Response:")
        print(json.dumps(response.json(), indent=2))
    else:
        print(f"Failed with status {response.status_code}")
        print(response.text)
except Exception as e:
    print(f"Error: {e}")
