requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
httpx
python-dotenv
python-multipart


